+++ scripts/utils.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import ffetch from './ffetch.js';
+
 import { BLOCKED_COUNTRIES } from './constants.js';
 
 const INDEX = '/query-index.json';
@@ -211,84 +211,6 @@ export const getGenericIndexData = (() => async () => {
   // Protected against callers modifying the objects
   return structuredClone(indexData);
 })();
-
-// export async function getIndexDataByPath(name) {
-function titleToName(name) {
-  return name.toLowerCase().replace(' ', '-');
-}
-
-const taxonomyEndpoint = '/tools/taxonomy.json';
-let taxonomyPromise;
-function fetchTaxonomy() {
-  if (!taxonomyPromise) {
-    taxonomyPromise = new Promise((resolve, reject) => {
-      (async () => {
-        try {
-          const taxonomyJson = await ffetch(taxonomyEndpoint).all();
-          const taxonomy = {};
-          let curType;
-          let l1;
-          taxonomyJson.forEach((row) => {
-            if (row.Type) {
-              curType = row.Type;
-              taxonomy[curType] = {
-                title: curType,
-                name: titleToName(curType),
-                path: titleToName(curType),
-                hide: row.hide,
-              };
-            }
-
-            if (row['Level 1']) {
-              l1 = row['Level 1'];
-              taxonomy[curType][l1] = {
-                title: l1,
-                name: titleToName(l1),
-                path: `${titleToName(curType)}/${titleToName(l1)}`,
-                hide: row.hide,
-              };
-            }
-          });
-          resolve(taxonomy);
-        } catch (e) {
-          reject(e);
-        }
-      })();
-    });
-  }
-
-  return taxonomyPromise;
-}
-
-const getDeepNestedObject = (obj, filter) => Object.entries(obj)
-  .reduce((acc, [key, value]) => {
-    let result = [];
-    if (key === filter) {
-      result = acc.concat(value);
-    } else if (typeof value === 'object') {
-      result = acc.concat(getDeepNestedObject(value, filter));
-    } else {
-      result = acc;
-    }
-    return result;
-  }, []);
-
-/**
- * Get the taxonomy of a hierarchical json object
- * @returns {Promise} the taxonomy
- */
-export function getTaxonomy() {
-  return fetchTaxonomy();
-}
-
-/**
- * Returns a taxonomy category as an array of objects
- * @param {*} category
- */
-export const getTaxonomyCategory = async (category) => {
-  const taxonomy = await getTaxonomy();
-  return getDeepNestedObject(taxonomy, category)[0];
-};
 
 // Function to create ellipsis
 function createEllipsis() {
